% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_selecter.R
\name{col_select}
\alias{col_select}
\title{Select Columns from a Dataframe}
\usage{
col_select(df, ret = c("df_select", "dplyr_code"), top_n = 100)
}
\arguments{
\item{df}{the \code{data.frame} you would like to subset.}

\item{ret}{should the subsetted data frame or the \code{dplyr}
code to make the subset be returned?}

\item{top_n}{for performance \code{col_select} only shows the first n
records, here you specify how many records you'd like to see. Default is 100.}
}
\description{
Shiny gadget that will show the data frame in a miniUI. Select the columns
that should be kept and hit Done.
}
\examples{
col_select(mtcars)
col_select(mtcars, 'dplyr_code')
}
\seealso{
\url{https://edwinth.github.io/blog/column-selecter/}
}
