% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-based_kappa.R
\name{get_kappa}
\alias{get_kappa}
\title{Get the Tree-Based Kappa}
\usage{
get_kappa(x, y, k = 10)
}
\arguments{
\item{x}{A numerical vector.}

\item{y}{A binary variable that can be coerced to factor.}

\item{k}{The number of folds to use in cross-validation.}
}
\value{
An object of class ggplot.
}
\description{
Split x and y into k folds. For each fold; train a univariate tree and get
the probalities for the hold-out. Then, obtain the optimal split and calculate
Cohen's kappa on y and y hat.
}
\seealso{
\url{https://edwinth.github.io/blog/tree-based-kappa/}
}
