% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-based_kappa.R
\name{gggini}
\alias{gggini}
\title{Make a gini plot}
\usage{
gggini(x, y)
}
\arguments{
\item{x}{A numerical vector.}

\item{y}{A binary variable that can be coerced to factor.}
}
\value{
An object of class ggplot.
}
\description{
Plots the ROC curve and also displays the univariate gini of \code{x}.
}
\seealso{
\url{https://edwinth.github.io/blog/tree-based-kappa/}
}
