% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmm.R
\name{ggmm}
\alias{ggmm}
\title{Create a Marimekko/Mosaic plot with ggplot2}
\usage{
ggmm(df, x, y, alpha_condition = 1 == 1, add_text = c(NULL, "n", "prop",
  "perc"), round_text = 2)
}
\arguments{
\item{df}{Data frame containing \code{x} and \code{y}.}

\item{x}{Bare name of the variable for the x-axis.}

\item{y}{Bare name of the variable for the y-axis and the fill.}

\item{alpha_condition}{All cells meeting this condition will have the alpha
set to 1. Cells not meeting the condition have the alpha set to 0.4.}

\item{add_text}{Add cell count, cell proportion of total, or cell
percentage of total.}

\item{round_text}{Nr of decimals, only applies to proportions and percentages.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Take two categorical variables and create a Marimekko/Mosaic plot from it.
}
\details{
The \code{alpha_condition} enables to highlight specific parts of
the plot. This is especially helpful for highlighting marginal distributions.
}
\examples{
ggmm(mtcars, cyl, vs)
ggmm(mtcars, cyl, vs, alpha_condition = vs == 1)
ggmm(mtcars, cyl, vs, alpha_condition = cyl == 6)
}
