% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-based_kappa.R
\name{predictor_plot}
\alias{predictor_plot}
\title{Make a Jitter Plot}
\usage{
predictor_plot(x, y)
}
\arguments{
\item{x}{A numerical vector.}

\item{y}{A binary variable that can be coerced to factor.}
}
\value{
An object of class ggplot.
}
\description{
Plot a numerical predictor agains a binary target.
}
\seealso{
\url{https://edwinth.github.io/blog/tree-based-kappa/}
}
