% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_to_wgs84.R
\name{rd_to_wgs84}
\alias{rd_to_wgs84}
\title{Convert RD to WGS84}
\source{
http://home.solcon.nl/pvanmanen/Download/Transformatieformules.pdf
Formula 6
}
\usage{
rd_to_wgs84(X, Y)
}
\arguments{
\item{X}{the X coordinate (mapped to longitude)}

\item{Y}{the Y coordinate (mapped to latitude)}
}
\value{
A double of length 2, the lat and the long.
}
\description{
Convert the (X, Y) of the Dutch standard RD (Rijksdriehoek) to the (lat, long)
of WGS84, as used in GPS.
}
