% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_plot.R
\name{tiles_pattern}
\alias{tiles_pattern}
\title{Make a random bathroom design}
\usage{
tiles_pattern(colors = c("darkblue", "cyan3", "lightblue1", "white"),
  nr_of_each_col = rep(40, 4), nr_height = 8, nr_width = 20,
  max_adjacent = 2)
}
\arguments{
\item{colors}{character vector with the colors to use.}

\item{nr_of_each_col}{prevalence of each color in \code{colors}, should
have same length as \code{colors}.}

\item{nr_height}{nr of tiles in vertical direction.}

\item{nr_width}{nr of tiles in horizontal directions.}

\item{max_adjacent}{maximum nr of adjacent tiles of the same color in
both directions.}
}
\value{
A \code{ggplot2} object with the pattern.
}
\description{
Function will generate a random pattern of tiles that can server
as a bathroom design.
}
\examples{
tiles_pattern()

}
\seealso{
\url{https://edwinth.github.io/blog/bathroom-with-r/}
}

