% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id.R
\name{unique_id}
\alias{unique_id}
\title{Check if the id columns are unique}
\usage{
unique_id(x, ...)
}
\arguments{
\item{x}{a data frame}

\item{...}{the unquoted column names of the id variables}
}
\value{
If the id variables are unique TRUE. Else, the rows in `x` for which the id
variables are not unique, arranged by id variables.
}
\description{
Check if the id columns are unique
}
\examples{
mtcars \%>\% unique_id(cyl, disp)
mtcars \%>\% unique_id(cyl, wt)
mtcars \%>\% unique_id(cyl, qsec)
}
